feat: implement wildcard routing for HTTP trigger
This commit implements the wildcard routing for the HTTP trigger.
Specifically, if a component declares the `/foo/...` route, all requests
that have that prefix route (`/foo`, or `/foo/bar`) should be routed to
that component.

Finally, if there are multiple components that attempt to handle the same
route, the last one in the configuration will be selected.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>