refactor key-value host component for per-component configuration
Previously, the set of available stores needed to be determined by the
time `KeyValueComponent` was created, which turned out to be too early
for some use cases.  In this commit, I've changed
`KeyValueComponent::new` to take a function which may be used to
create a `StoreManager` (formerly known as `Impl`) for any number of
applications.  That `StoreManager`, in turn, may be used to open any
number of `Store`s (formerly known as `ImplStore`s).

Note that I've used `Arc<dyn _>` rather than `Box<dyn _>` to represent
trait objects in order to facilitate connection reuse where
appropriate, among both `Store`s and `StoreManager`s.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>