move Rust and Go SDKs to their own repos (#2289)
* move Rust and Go SDKs to their own repos

These SDKs and their examples have moved to
https://github.com/fermyon/spin-rust-sdk and
https://github.com/fermyon/spin-go-sdk, respectively.  The plan is to begin
releasing them on their own cadence, independent of Spin releases.  This will
allow us to make breaking changes in semver-compatible major releases without
tying such releases to a Spin major release, e.g. to target WASI 0.2.0.

Note that I've update the integration tests which use the Rust SDK to use the
latest stable release (2.2.0) instead of a path-based dependency.
Alternatively, we could point it at the `main` branch of the `spin-rust-sdk` Git
repo, but that will require close coordination once we start making breaking
changes to the API in that repo.

We may also want to move the Rust and Go templates to the new repos at some
point.  I've left them where they are for the time being.

Finally, I've removed the SDK-specific bits from release-process.md and the
GitHub workflows.  We'll need to move those things to their respective SDK
repos.  Reproducing the automated Go SDK tagging that feeds into the templates
may require moving at least the Go templates as noted above.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* restore accidentally-removed lines from release-process.md

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* update SDK examples links to point to `stable` branches

We'll update the `stable` branch of the applicable repo each time we make a new
stable release of any of the SDKs so that these links always point to the
examples for the latest stable release rather than `main`.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* re-add (updated version of) `push-templates-tag` to release.yml

Unlike the original version, this one does not depend on the Go SDK automation
and instead derives the template tag name from the Spin version tag name.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* re-add `build-and-sign` as dependency of `push-templates-tag`

This ensures that we don't tag a broken build.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

---------

Signed-off-by: Joel Dice <joel.dice@fermyon.com>